+++ rust/examples/echo.rs
@@ -187,6 +187,7 @@ impl EchoConnection {
                 ip: IpAddr::V4(Ipv4Addr::LOCALHOST),
                 announced_address: None,
                 port: None,
+                port_range: None,
                 flags: None,
                 send_buffer_size: None,
                 recv_buffer_size: None,