+++ rust/tests/integration/worker.rs
@@ -42,7 +42,7 @@ fn create_worker_succeeds() {
 
                 settings.log_level = WorkerLogLevel::Debug;
                 settings.log_tags = vec![WorkerLogTag::Info];
-                settings.rtc_ports_range = 0..=9999;
+                settings.rtc_port_range = 0..=9999;
                 settings.dtls_files = Some(WorkerDtlsFiles {
                     certificate: "tests/integration/data/dtls-cert.pem".into(),
                     private_key: "tests/integration/data/dtls-key.pem".into(),
@@ -79,7 +79,7 @@ fn create_worker_wrong_settings() {
                     // Intentionally incorrect range
                     #[allow(clippy::reversed_empty_ranges)]
                     {
-                        settings.rtc_ports_range = 1000..=999;
+                        settings.rtc_port_range = 1000..=999;
                     }
 
                     settings