+++ worker/fuzzer/src/fuzzer.cpp
@@ -9,8 +9,14 @@
 #include "LogLevel.hpp"
 #include "Settings.hpp"
 #include "Utils.hpp"
+#include "RTC/Codecs/FuzzerH264.hpp"
+#include "RTC/Codecs/FuzzerH264_SVC.hpp"
+#include "RTC/Codecs/FuzzerOpus.hpp"
+#include "RTC/Codecs/FuzzerVP8.hpp"
+#include "RTC/Codecs/FuzzerVP9.hpp"
 #include "RTC/DtlsTransport.hpp"
 #include "RTC/FuzzerDtlsTransport.hpp"
+#include "RTC/FuzzerRateCalculator.hpp"
 #include "RTC/FuzzerRtpPacket.hpp"
 #include "RTC/FuzzerRtpRetransmissionBuffer.hpp"
 #include "RTC/FuzzerRtpStreamSend.hpp"
@@ -23,13 +29,16 @@
 #include <stddef.h>
 #include <stdint.h>
 
-bool fuzzStun  = false;
-bool fuzzDtls  = false;
-bool fuzzRtp   = false;
-bool fuzzRtcp  = false;
-bool fuzzUtils = false;
+bool fuzzStun   = false;
+bool fuzzDtls   = false;
+bool fuzzRtp    = false;
+bool fuzzRtcp   = false;
+bool fuzzCodecs = false;
+bool fuzzUtils  = false;
 
-int Init();
+// This Init() function must be declared static, otherwise linking will fail if
+// another source file defines same non static Init() function.
+static int Init();
 
 extern "C" int LLVMFuzzerTestOneInput(const uint8_t* data, size_t len)
 {
@@ -55,11 +64,21 @@ extern "C" int LLVMFuzzerTestOneInput(const uint8_t* data, size_t len)
 		Fuzzer::RTC::RtpStreamSend::Fuzz(data, len);
 		Fuzzer::RTC::RtpRetransmissionBuffer::Fuzz(data, len);
 		Fuzzer::RTC::SeqManager::Fuzz(data, len);
+		Fuzzer::RTC::RateCalculator::Fuzz(data, len);
 	}
 
 	if (fuzzRtcp)
 	{
 		Fuzzer::RTC::RTCP::Packet::Fuzz(data, len);
+	}
+
+	if (fuzzCodecs)
+	{
+		Fuzzer::RTC::Codecs::Opus::Fuzz(data, len);
+		Fuzzer::RTC::Codecs::VP8::Fuzz(data, len);
+		Fuzzer::RTC::Codecs::VP9::Fuzz(data, len);
+		Fuzzer::RTC::Codecs::H264::Fuzz(data, len);
+		Fuzzer::RTC::Codecs::H264_SVC::Fuzz(data, len);
 	}
 
 	if (fuzzUtils)
@@ -118,21 +137,28 @@ int Init()
 
 		fuzzRtcp = true;
 	}
+	if (std::getenv("MS_FUZZ_CODECS") && std::string(std::getenv("MS_FUZZ_CODECS")) == "1")
+	{
+		std::cout << "[fuzzer] codecs fuzzer enabled" << std::endl;
+
+		fuzzCodecs = true;
+	}
 	if (std::getenv("MS_FUZZ_UTILS") && std::string(std::getenv("MS_FUZZ_UTILS")) == "1")
 	{
 		std::cout << "[fuzzer] Utils fuzzer enabled" << std::endl;
 
 		fuzzUtils = true;
 	}
-	if (!fuzzStun && !fuzzDtls && !fuzzRtcp && !fuzzRtp && !fuzzUtils)
+	if (!fuzzStun && !fuzzDtls && !fuzzRtp && !fuzzRtcp && !fuzzCodecs && !fuzzUtils)
 	{
 		std::cout << "[fuzzer] all fuzzers enabled" << std::endl;
 
-		fuzzStun  = true;
-		fuzzDtls  = true;
-		fuzzRtp   = true;
-		fuzzRtcp  = true;
-		fuzzUtils = true;
+		fuzzStun   = true;
+		fuzzDtls   = true;
+		fuzzRtp    = true;
+		fuzzRtcp   = true;
+		fuzzCodecs = true;
+		fuzzUtils  = true;
 	}
 
 	Settings::configuration.logLevel = logLevel;